V1

+++ resources/views/livewire/d-p-a/kegiatan/table.blade.php
@@ -103,10 +103,12 @@
                                     class="btn btn-info btn-icon">
                                     <i class="ik ik-corner-down-right"></i>
                                 </a>
-                                <button class="btn btn-danger btn-icon ml-2"
-                                    onclick="return confirm('Ingin menghapus Kegiatan ini?')"
-                                    wire:click.prevent='destroyKegiatan("{{ $kegiatan->uuid }}")'><i
-                                        class="ik ik-trash-2"></i></button>
+                                @if ($kegiatan->subkegiatan->count() == 0)
+                                    <button class="btn btn-danger btn-icon ml-2"
+                                        onclick="return confirm('Ingin menghapus Kegiatan ini?')"
+                                        wire:click.prevent='destroyKegiatan("{{ $kegiatan->uuid }}")'><i
+                                            class="ik ik-trash-2"></i></button>
+                                @endif
                             </div>
                         </td>
                         <td class="text-center">
